causal_test_result + new dataclass for test values
